Fix background stream list updates
If no other thread has obtained the skip_cond_lock, i.e. if no thread
serving a request is current blocking on the condition variable, i.e. if
no requests are currently being served, set skip_cond to true so that if
it is still true that no connections are being served when the
background thread goes to update the stream list, it skips waiting on
the condition variable.

Otherwise, the background thread just blocks forever which is what was
happening before this change :^)