Mega clean-up
Introduce the `GUARD` macro to reduce line noise and also guarantee each
mutex lock is accompanied by an unlock.

Use a much simpler strategy for serving the streams list over HTTP to
replace the previous condition variable based solution. The new strategy
is simpler and maintains the desirable property that only one thread is
re-computing the sorted stream list at a time.

Numerous other small fixes. Hopefully this improves the maintainability
of the software in the future.